backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
uvicorn==0.24.0
fastapi==0.100.1

# LLM and Embeddings (OpenAI)
openai>=1.6.1
langchain>=0.1.20
langchain-core>=0.1.30
langchain-community>=0.0.15
langchain-openai>=0.0.2

# Vector Database
faiss-cpu==1.7.4

# Document Processing
pypdf==3.16.0
python-docx==0.8.11
markdown==3.4.1

# Others
requests>=2.31.0
gunicorn==21.2.0
